% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{poisDoubleSamp}
\alias{package-poisDoubleSamp}
\alias{poisDoubleSamp}
\alias{poisDoubleSamp-package}
\title{poisDoubleSamp : Confidence intervals with Poisson double sampling}
\description{
Functions to create confidence intervals for ratios of Poisson rates under misclassification using double sampling.
}

