% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/help.r
\docType{package}
\name{poisDoubleSamp}
\alias{package-poisDoubleSamp}
\alias{poisDoubleSamp}
\alias{poisDoubleSamp-package}
\title{poisDoubleSamp : Confidence intervals with Poisson double sampling}
\description{
Functions to create confidence intervals for ratios of Poisson rates under misclassification using double sampling.
}

